# [WIP] Fix type conversion warnings on Windows