--- a/src/generic/json_structural_indexer.h
+++ b/src/generic/json_structural_indexer.h
@@ -22,10 +22,11 @@ public:
     // it helps tremendously.
     if (bits == 0)
         return;
-    uint32_t cnt = count_ones(bits);
+    auto cnt = count_ones(bits);
 
     // Do the first 8 all together
-    for (int i=0; i<8; i++) {
+    decltype(cnt) i; // Make i the same type as cnt (i.e. result type of count_ones)
+    for (i=0; i<8; i++) {
       this->tail[i] = idx + trailing_zeroes(bits);
       bits = clear_lowest_bit(bits);
     }
@@ -33,7 +34,7 @@ public:
     // Do the next 8 all together (we hope in most cases it won't happen at all
     // and the branch is easily predicted).
     if (unlikely(cnt > 8)) {
-      for (int i=8; i<16; i++) {
+      for (i=8; i<16; i++) {
         this->tail[i] = idx + trailing_zeroes(bits);
         bits = clear_lowest_bit(bits);
       }
@@ -42,7 +43,7 @@ public:
       // branch mispredictions here. 16+ structurals per block means either punctuation ({} [] , :)
       // or the start of a value ("abc" true 123) every four characters.
       if (unlikely(cnt > 16)) {
-        uint32_t i = 16;
+        i = 16;
         do {
           this->tail[i] = idx + trailing_zeroes(bits);
           bits = clear_lowest_bit(bits);
@@ -57,15 +58,15 @@ public:
 
 class json_structural_indexer {
 public:
-  template<size_t STEP_SIZE>
-  static error_code index(const uint8_t *buf, size_t len, parser &parser, bool streaming) noexcept;
+  template<uint32_t STEP_SIZE>
+  static error_code index(const uint8_t *buf, uint32_t len, parser &parser, bool streaming) noexcept;
 
 private:
   really_inline json_structural_indexer(uint32_t *structural_indexes) : indexer{structural_indexes} {}
-  template<size_t STEP_SIZE>
+  template<uint32_t STEP_SIZE>
   really_inline void step(const uint8_t *block, buf_block_reader<STEP_SIZE> &reader) noexcept;
-  really_inline void next(simd::simd8x64<uint8_t> in, json_block block, size_t idx);
-  really_inline error_code finish(parser &parser, size_t idx, size_t len, bool streaming);
+  really_inline void next(simd::simd8x64<uint8_t> in, json_block block, uint32_t idx);
+  really_inline error_code finish(parser &parser, uint32_t idx, uint32_t len, bool streaming);
 
   json_scanner scanner;
   utf8_checker checker{};
@@ -74,7 +75,7 @@ private:
   uint64_t unescaped_chars_error = 0;
 };
 
-really_inline void json_structural_indexer::next(simd::simd8x64<uint8_t> in, json_block block, size_t idx) {
+really_inline void json_structural_indexer::next(simd::simd8x64<uint8_t> in, json_block block, uint32_t idx) {
   uint64_t unescaped = in.lteq(0x1F);
   checker.check_next_input(in);
   indexer.write(idx-64, prev_structurals); // Output *last* iteration's structurals to the parser
@@ -82,7 +83,7 @@ really_inline void json_structural_indexer::next(simd::simd8x64<uint8_t> in, jso
   unescaped_chars_error |= block.non_quote_inside_string(unescaped);
 }
 
-really_inline error_code json_structural_indexer::finish(parser &parser, size_t idx, size_t len, bool streaming) {
+really_inline error_code json_structural_indexer::finish(parser &parser, uint32_t idx, uint32_t len, bool streaming) {
   // Write out the final iteration's structurals
   indexer.write(idx-64, prev_structurals);
 
@@ -93,7 +94,7 @@ really_inline error_code json_structural_indexer::finish(parser &parser, size_t 
     return UNESCAPED_CHARS;
   }
 
-  parser.n_structural_indexes = indexer.tail - parser.structural_indexes.get();
+  parser.n_structural_indexes = static_cast<uint32_t>(indexer.tail - parser.structural_indexes.get());
   /* a valid JSON file cannot have zero structural indexes - we should have
    * found something */
   if (unlikely(parser.n_structural_indexes == 0u)) {
@@ -153,8 +154,8 @@ really_inline void json_structural_indexer::step<64>(const uint8_t *block, buf_b
 // Setting the streaming parameter to true allows the find_structural_bits to tolerate unclosed strings.
 // The caller should still ensure that the input is valid UTF-8. If you are processing substrings,
 // you may want to call on a function like trimmed_length_safe_utf8.
-template<size_t STEP_SIZE>
-error_code json_structural_indexer::index(const uint8_t *buf, size_t len, parser &parser, bool streaming) noexcept {
+template<uint32_t STEP_SIZE>
+error_code json_structural_indexer::index(const uint8_t *buf, uint32_t len, parser &parser, bool streaming) noexcept {
   if (unlikely(len > parser.capacity())) { return CAPACITY; }
 
   buf_block_reader<STEP_SIZE> reader(buf, len);