--- a/src/haswell/stage1_find_marks.h
+++ b/src/haswell/stage1_find_marks.h
@@ -72,7 +72,7 @@ WARN_UNUSED error_code implementation::minify(const uint8_t *buf, size_t len, ui
 #include "generic/utf8_lookup2_algorithm.h"
 #include "generic/json_structural_indexer.h"
 WARN_UNUSED error_code implementation::stage1(const uint8_t *buf, size_t len, parser &parser, bool streaming) const noexcept {
-  return haswell::stage1::json_structural_indexer::index<128>(buf, len, parser, streaming);
+  return haswell::stage1::json_structural_indexer::index<128>(buf, static_cast<uint32_t>(len), parser, streaming);
 }
 
 } // namespace haswell