Paying down the Windows warning debt.

Also moved the flags into simdjson-flags.cmake, something I've wanted to do. The general recommendation for cmake projects is to have the top level cmake mostly include other stuff.