@jkeiser 2020-04-20
Yep, this causes about a 5% stage 1 performance degradation and 1% overall. I'll back off the datatype changes until it recovers.