@lemire 2020-04-20
> Once I fixed the warnings emitted by clang on the other end, the performance righted itself.

Can you elaborate?