@jkeiser 2020-04-20
> Can you elaborate?

When I changed types, Windows stopped warning, but g++7 (sorry, thought it was clang) picked up some new warnings, particularly signed vs. unsigned comparisons. I changed a couple more types at that point. It may or may not be those changes that fixed perf, but perf does appear to be fixed on my end at least.