@lemire 2020-04-20
> When I changed types, Windows stopped warning, but g++7 (sorry, thought it was clang) picked up some new warnings, particularly signed vs. unsigned comparisons. I changed a couple more types at that point.

So Visual Studio is fine with == when comparing signed and unsigned... but will complain with inequalities (C4018). Clang should complain in both cases.

I can't imagine how it could massively impact performance... but there you have it.