@jkeiser 2020-04-20
> I can't imagine how it could massively impact performance... but there you have it.

I'm chalking it up to volatility; but if you wanted to check the perf on another machine, I wouldn't mind the confirmation :)