@lemire 2020-04-20
It looks like there is significant performance regression on westmere...

This PR:
```
|- Stage 1
|    Speed        :  21.3537 ns per block ( 55.80%) -   0.3337 ns per byte -   3.8130 ns per structural -    2.997 GB/s
|    Cycles       :  78.8400 per block    ( 56.12%) -   1.2319 per byte    -  14.0780 per structural    -    3.692 GHz est. frequency
|    Instructions : 293.2226 per block    ( 60.35%) -   4.5819 per byte    -  52.3591 per structural    -    3.719 per cycle
|    Misses       :    1411 branch misses ( 46.17%) - 0 cache misses (  0.00%) - 28063.00 cache references
```


Master:
```
|- Stage 1
|    Speed        :  20.4545 ns per block ( 53.62%) -   0.3196 ns per byte -   3.6524 ns per structural -    3.129 GB/s
|    Cycles       :  75.5329 per block    ( 53.94%) -   1.1803 per byte    -  13.4875 per structural    -    3.693 GHz est. frequency
|    Instructions : 287.3537 per block    ( 59.86%) -   4.4902 per byte    -  51.3111 per structural    -    3.804 per cycle
|    Misses       :    1180 branch misses ( 38.22%) - 15 cache misses ( 34.95%) - 30217.00 cache references
```

I am using twitter.json.